pyproject.toml
[tool.poetry]
name = "velo-action"
version = "0.1.0"
description = "Github Action used to generate velo deploy version and trigger a release."
authors = ["Cloud-platform team <team-cloud-platform@oda.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
semantic-version = "^2.9"
appdirs = "^1.4"
google = "^3.0"
google-auth = "^2.6.0"
google-cloud-storage = "^2.1"
google-cloud-secret-manager = "^2.9"
pydantic = "^1.8.2"
opentelemetry-sdk = "^1.6.2"
opentelemetry-exporter-jaeger = "^1.6.2"
opentelemetry-api = "^1.6.2"
opentelemetry-exporter-otlp = "^1.6.2"
PyYAML = "^6.0"
types-PyYAML = "^6.0.3"
loguru = "^0.6.0"
semver = "^2.13.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0"
black = "^22.1.0"
flake8 = "^4.0"
flake8-bugbear = "^22.1"
mypy = "^0.931"
pylint = "^2.12"
pytest-mock = "^3.6"
docker-image-size-limit = "^0.4.1"
types-requests = "^2.26.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = ['py310']
safe = true
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.venv
  | _build
  | build
  | dist
  | migrations
)/
'''
